Remove `anchors()` from the public JSON Schema interface
