+++ src/core/jsonschema/CMakeLists.txt
@@ -4,11 +4,11 @@ include(./official_resolver.cmake)
 
 sourcemeta_library(NAMESPACE sourcemeta PROJECT core NAME jsonschema
   FOLDER "Core/JSON Schema"
-  PRIVATE_HEADERS anchor.h bundle.h resolver.h
+  PRIVATE_HEADERS bundle.h resolver.h
     walker.h reference.h frame.h error.h unevaluated.h
     keywords.h transform.h
   SOURCES jsonschema.cc default_walker.cc frame.cc
-    anchor.cc resolver.cc walker.cc bundle.cc
+    resolver.cc walker.cc bundle.cc
     unevaluated.cc relativize.cc unidentify.cc
     transform_rule.cc transformer.cc
     "${CMAKE_CURRENT_BINARY_DIR}/official_resolver.cc")
