--- src/core/jsonschema/anchor.cc
@@ -1,101 +1,0 @@
-#include <sourcemeta/core/jsonschema.h>
-#include <sourcemeta/core/uri.h>
-
-#include <cassert> // assert
-#include <utility> // std::move
-
-namespace sourcemeta::core {
-
-auto anchors(const JSON &schema, const SchemaResolver &resolver,
-             const std::optional<std::string> &default_dialect)
-    -> std::map<std::string, AnchorType> {
-  const std::map<std::string, bool> vocabularies{
-      sourcemeta::core::vocabularies(schema, resolver, default_dialect)};
-  return anchors(schema, vocabularies);
-}
-
-auto anchors(const JSON &schema,
-             const std::map<std::string, bool> &vocabularies)
-    -> std::map<std::string, AnchorType> {
-  std::map<std::string, AnchorType> result;
-
-  // 2020-12
-  if (schema.is_object() &&
-      vocabularies.contains(
-          "https://json-schema.org/draft/2020-12/vocab/core")) {
-    if (schema.defines("$dynamicAnchor")) {
-      const auto &anchor{schema.at("$dynamicAnchor")};
-      assert(anchor.is_string());
-      result.insert({anchor.to_string(), AnchorType::Dynamic});
-    }
-
-    if (schema.defines("$anchor")) {
-      const auto &anchor{schema.at("$anchor")};
-      assert(anchor.is_string());
-      const auto anchor_string{anchor.to_string()};
-      const auto success = result.insert({anchor_string, AnchorType::Static});
-      assert(success.second || result.contains(anchor_string));
-      if (!success.second) {
-        result[anchor_string] = AnchorType::All;
-      }
-    }
-  }
-
-  // 2019-09
-  if (schema.is_object() &&
-      vocabularies.contains(
-          "https://json-schema.org/draft/2019-09/vocab/core")) {
-    if (schema.defines("$recursiveAnchor")) {
-      const auto &anchor{schema.at("$recursiveAnchor")};
-      assert(anchor.is_boolean());
-      if (anchor.to_boolean()) {
-        // We store a 2019-09 recursive anchor as an empty anchor
-        result.insert({"", AnchorType::Dynamic});
-      }
-    }
-
-    if (schema.defines("$anchor")) {
-      const auto &anchor{schema.at("$anchor")};
-      assert(anchor.is_string());
-      const auto anchor_string{anchor.to_string()};
-      const auto success = result.insert({anchor_string, AnchorType::Static});
-      assert(success.second || result.contains(anchor_string));
-      if (!success.second) {
-        result[anchor_string] = AnchorType::All;
-      }
-    }
-  }
-
-  // Draft 7 and 6
-  // Old `$id` anchor form
-  if (schema.is_object() &&
-      (vocabularies.contains("http://json-schema.org/draft-07/schema#") ||
-       vocabularies.contains("http://json-schema.org/draft-06/schema#"))) {
-    if (schema.defines("$id")) {
-      assert(schema.at("$id").is_string());
-      const URI identifier(schema.at("$id").to_string());
-      if (identifier.is_fragment_only()) {
-        result.insert(
-            {std::string{identifier.fragment().value()}, AnchorType::Static});
-      }
-    }
-  }
-
-  // Draft 4
-  // Old `id` anchor form
-  if (schema.is_object() &&
-      vocabularies.contains("http://json-schema.org/draft-04/schema#")) {
-    if (schema.defines("id")) {
-      assert(schema.at("id").is_string());
-      const URI identifier(schema.at("id").to_string());
-      if (identifier.is_fragment_only()) {
-        result.insert(
-            {std::string{identifier.fragment().value()}, AnchorType::Static});
-      }
-    }
-  }
-
-  return result;
-}
-
-} // namespace sourcemeta::core
